+++ src/Models/BlogEtcUploadedPhoto.php
@@ -2,17 +2,10 @@
 
 namespace WebDevEtc\BlogEtc\Models;
 
-use Illuminate\Database\Eloquent\Model;
-
-class BlogEtcUploadedPhoto extends Model
+/**
+ * legacy class for backwards compatibility.
+ * @deprecated
+ */
+class BlogEtcUploadedPhoto extends UploadedPhoto
 {
-    public $table = 'blog_etc_uploaded_photos';
-    public $casts = [
-        'uploaded_images' => 'array',
-    ];
-    public $fillable = [
-        'image_title',
-        'uploader_id',
-        'source', 'uploaded_images',
-    ];
 }
